move causalml generate method to causalmlpipelineextension
