+++ src/Microsoft.ML.GenAI.Core/Extension/ModuleExtension.cs
@@ -160,41 +160,6 @@ internal static class ModuleExtension
         return deviceMap;
     }
 
-    public static string? Generate(
-        this CausalLMPipeline pipeline,
-        string prompt,
-        int maxLen = 128,
-        float temperature = 0.7f,
-        float topP = 0.9f,
-        string[]? stopSequences = null,
-        int eosId = 0,
-        string device = "cpu",
-        bool bos = true,
-        bool eos = false,
-        bool echo = false)
-    {
-        using var newScope = NewDisposeScope();
-        var inputIds = pipeline.Tokenizer.EncodeToIds(prompt);
-        var inputTensor = torch.tensor(inputIds.ToArray(), dtype: ScalarType.Int64, device: device).unsqueeze(0);
-        var attentionMask = torch.ones_like(inputTensor);
-
-        // set up stop token ids
-        // stop token ids: [[eosId], [stopSequence1], [stopSequence2], ...]
-        // when causal language model generates tokens, it will stop when it generates any token in stopSequences
-        List<int[]> stopTokenIds = [[eosId]];
-        if (stopSequences != null)
-        {
-            stopTokenIds.AddRange(stopSequences.Select(x => pipeline.Tokenizer.EncodeToIds(x).ToArray()));
-        }
-
-        (var token, var _) = pipeline.Generate(inputTensor, attentionMask, temperature: temperature, maxLen: maxLen, topP: topP, stopTokenSequence: stopTokenIds.ToArray(), echo: echo);
-
-        var tokenIds = token[0].to_type(ScalarType.Int32).data<int>().ToArray();
-
-        return pipeline.Tokenizer.Decode(tokenIds);
-    }
-
-
     public static string Peek(this nn.Module model)
     {
         var sb = new StringBuilder();
